Decouples popup from controls overlay to reuse in passive pickups
